LUM-9537 add notes path

+++ src/main/java/com/blackbaud/deployment/core/domain/ReleasePlanRepository.java
@@ -2,7 +2,7 @@ package com.blackbaud.deployment.core.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReleasePlanRepository extends CrudRepository<ReleasePlanEntity, Integer> {
+public interface ReleasePlanRepository extends CrudRepository<ReleasePlanEntity, Long> {
 
     ReleasePlanEntity findByClosedNull();
 }
